requirements.txt
djangorestframework==3.12.2
django==3.1.3
